Enhanced fetching all and enrolled courses in teachers dashboard component to fix rendering issues on reload or refresh
